backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
Pillow  # Per gestire le immagini (ImageField nel CMS)
reportlab
